% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bdeb_plots}
\alias{bdeb_plots}
\title{Plot Methods for BDEB Objects}
\description{
Visualisation methods for BDEB fits, posterior predictive checks,
and derived quantities using ggplot2.  All plots return ggplot2
objects that can be further customised.
}
