% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{build_stan_data_debtox}
\alias{build_stan_data_debtox}
\title{Build Stan Data List for DEBtox}
\usage{
build_stan_data_debtox(data, priors)
}
\arguments{
\item{data}{A \code{bdeb_data} object with concentration info.}

\item{priors}{A list of \code{bdeb_prior} objects.}
}
\value{
Named list suitable for Stan.
}
\description{
Build Stan Data List for DEBtox
}
\keyword{internal}
