% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{build_stan_data_individual}
\alias{build_stan_data_individual}
\title{Build Stan Data List for Individual Growth}
\usage{
build_stan_data_individual(data, priors)
}
\arguments{
\item{data}{A \code{bdeb_data} object.}

\item{priors}{A list of \code{bdeb_prior} objects.}
}
\value{
Named list suitable for Stan.
}
\description{
Build Stan Data List for Individual Growth
}
\keyword{internal}
