% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eisenia_cd}
\alias{eisenia_cd}
\title{Eisenia andrei Cadmium Toxicity Data (Van Gestel 1991)}
\format{
A data frame with 35 rows and 4 columns:
\describe{
\item{id}{Concentration group identifier (1--5)}
\item{time}{Time in days (0--85)}
\item{length}{Structural length in cm}
\item{concentration}{Cadmium concentration (mg Cd/kg soil)}
}
}
\source{
EGrowth database, curve IDs gr0119--gr0123.
\url{https://github.com/JeromeMathieuEcology/EGrowth}
}
\usage{
eisenia_cd
}
\description{
Real experimental growth data for \emph{Eisenia andrei} exposed to
cadmium in natural soil at 23 °C from Van Gestel et al. (1991),
obtained via the EGrowth database (Mathieu, 2018).  Five
concentration groups (0, 10, 32, 100, 320 mg Cd/kg) with 7 time
points each over 85 days.  Group-mean body mass was converted to
structural length via \eqn{L = (W / d_V)^{1/3}} with
\eqn{d_V = 1.05} g cm\eqn{^{-3}}.
}
\examples{
data(eisenia_cd)
plot(eisenia_cd$time, eisenia_cd$length,
     col = as.factor(eisenia_cd$concentration),
     xlab = "Days", ylab = "Structural length (cm)")
}
\references{
Van Gestel, C.A.M., Van Dis, W.A., Dirven-Van Breemen, E.M.,
Sparenburg, P.M. and Baerselman, R. (1991). Influence of cadmium,
copper, and pentachlorophenol on growth and sexual development of
\emph{Eisenia andrei} (Oligochaeta; Annelida). \emph{Biology and Fertility
of Soils}, 12, 117--121. \doi{10.1007/BF00341486}
}
\keyword{datasets}
