% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eisenia_neuhauser}
\alias{eisenia_neuhauser}
\title{Eisenia fetida Growth Data (Neuhauser 1980)}
\format{
A data frame with 37 rows and 3 columns:
\describe{
\item{id}{Individual/group identifier (always 1)}
\item{time}{Time in days since start (1--250)}
\item{length}{Structural length in cm (0.20--1.31)}
}
}
\source{
EGrowth database, curve ID gr0226.
\url{https://github.com/JeromeMathieuEcology/EGrowth}
}
\usage{
eisenia_neuhauser
}
\description{
Real experimental growth data for \emph{Eisenia fetida} on activated
sludge at 25 °C from Neuhauser, Hartenstein & Kaplan (1980),
obtained via the EGrowth database (Mathieu, 2018).  The dataset
comprises 37 group-mean body mass measurements (20 worms per
replicate) over 250 days, from hatchling (~8 mg) to adult
(~2350 mg).  Wet mass was converted to structural length via
\eqn{L = (W / d_V)^{1/3}} with tissue density
\eqn{d_V = 1.05} g cm\eqn{^{-3}}.
}
\examples{
data(eisenia_neuhauser)
plot(eisenia_neuhauser$time, eisenia_neuhauser$length,
     xlab = "Days", ylab = "Structural length (cm)")
}
\references{
Neuhauser, E.F., Hartenstein, R. and Kaplan, D.L. (1980). Growth
of the earthworm \emph{Eisenia foetida} in relation to population density
and food rationing. \emph{Oikos}, 35, 93--98.
\doi{10.2307/3544730}

Mathieu, J. (2018). EGrowth: a global database on intraspecific
body growth variability in earthworm. \emph{Soil Biology and
Biochemistry}, 122, 71--80. \doi{10.1016/j.soilbio.2018.04.004}
}
\keyword{datasets}
