% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{observation_to_stan_data}
\alias{observation_to_stan_data}
\title{Encode observation model flags for Stan}
\usage{
observation_to_stan_data(observation)
}
\arguments{
\item{observation}{Named list of bdeb_obs objects.}
}
\value{
Named list with obs_growth, obs_nu, obs_repro.
}
\description{
Encode observation model flags for Stan
}
\keyword{internal}
