% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{print.bdeb_data}
\alias{print.bdeb_data}
\title{Print a BDEB Data Object}
\usage{
\method{print}{bdeb_data}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=bdeb_data]{bdeb_data()}} object.}

\item{...}{Ignored.}
}
\value{
The input object, invisibly.
}
\description{
Print a BDEB Data Object
}
