% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.bdeb_fit}
\alias{print.bdeb_fit}
\title{Print a BDEB Fit}
\usage{
\method{print}{bdeb_fit}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=bdeb_fit]{bdeb_fit()}} object.}

\item{...}{Ignored.}
}
\value{
The input object, invisibly.
}
\description{
Print a BDEB Fit
}
