% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{print.bdeb_model}
\alias{print.bdeb_model}
\title{Print a BDEB Model Specification}
\usage{
\method{print}{bdeb_model}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=bdeb_model]{bdeb_model()}} object.}

\item{...}{Ignored.}
}
\value{
The input object, invisibly.
}
\description{
Print a BDEB Model Specification
}
