% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc.R
\name{print.bdeb_ppc}
\alias{print.bdeb_ppc}
\title{Print a BDEB Posterior Predictive Check}
\usage{
\method{print}{bdeb_ppc}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=bdeb_ppc]{bdeb_ppc()}} object.}

\item{...}{Ignored.}
}
\value{
The input object, invisibly.
}
\description{
Print a BDEB Posterior Predictive Check
}
