% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{stan_file}
\alias{stan_file}
\title{Get path to a bundled Stan model file}
\usage{
stan_file(model_name)
}
\arguments{
\item{model_name}{Name of the Stan model (without .stan extension).}
}
\value{
Full path to the \code{.stan} file.
}
\description{
Get path to a bundled Stan model file
}
\keyword{internal}
