% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{summary.bdeb_data}
\alias{summary.bdeb_data}
\alias{print.summary.bdeb_data}
\title{Summary of a BDEB Data Object}
\usage{
\method{summary}{bdeb_data}(object, ...)

\method{print}{summary.bdeb_data}(x, ...)
}
\arguments{
\item{object}{A \code{\link[=bdeb_data]{bdeb_data()}} object.}

\item{...}{Ignored.}

\item{x}{A \code{summary.bdeb_data} object.}
}
\value{
An object of class \code{summary.bdeb_data} (a list).
}
\description{
Returns a compact list of summary statistics describing a
\code{\link[=bdeb_data]{bdeb_data()}} object: number of observations, number of unique
individuals, time range, presence of growth/reproduction endpoints,
the functional response, and (for DEBtox data) the unique
concentration levels.
}
\examples{
df <- data.frame(id = 1, time = 0:5, length = seq(0.1, 0.6, by = 0.1))
dat <- bdeb_data(growth = df)
summary(dat)
}
