% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{temperature_to_stan_data}
\alias{temperature_to_stan_data}
\title{Encode temperature correction for Stan}
\usage{
temperature_to_stan_data(temperature)
}
\arguments{
\item{temperature}{NULL or list with T_obs, T_ref, T_A.}
}
\value{
Named list with has_temperature, T_obs, T_ref, T_A.
}
\description{
Encode temperature correction for Stan
}
\keyword{internal}
