% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{vcov.bdeb_fit}
\alias{vcov.bdeb_fit}
\title{Posterior Covariance Matrix of BDEB Model Parameters}
\usage{
\method{vcov}{bdeb_fit}(object, ...)
}
\arguments{
\item{object}{A \code{\link[=bdeb_fit]{bdeb_fit()}} object.}

\item{...}{Ignored.}
}
\value{
A symmetric numeric matrix with model parameters on rows
and columns.
}
\description{
Computes the empirical covariance matrix of the posterior draws for
all model parameters.  Bayesian counterpart of \code{\link[stats:vcov]{stats::vcov()}}.
}
