% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{build_stan_data_growth_repro}
\alias{build_stan_data_growth_repro}
\title{Build Stan Data List for Growth + Reproduction}
\usage{
build_stan_data_growth_repro(
  data,
  priors,
  temperature = NULL,
  observation = NULL
)
}
\arguments{
\item{data}{A \code{bdeb_data} object with growth and reproduction.}

\item{priors}{A list of \code{bdeb_prior} objects.}
}
\value{
Named list suitable for Stan.
}
\description{
Build Stan Data List for Growth + Reproduction
}
\keyword{internal}
