% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{build_stan_data_individual}
\alias{build_stan_data_individual}
\title{Build Stan Data List for Individual Growth}
\usage{
build_stan_data_individual(
  data,
  priors,
  temperature = NULL,
  observation = NULL
)
}
\arguments{
\item{data}{A \code{bdeb_data} object.}

\item{priors}{A list of \code{bdeb_prior} objects.}

\item{temperature}{NULL or list with T_obs, T_ref, T_A.}
}
\value{
Named list suitable for Stan.
}
\description{
Build Stan Data List for Individual Growth
}
\keyword{internal}
