% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bdeb_ppc}
\alias{plot.bdeb_ppc}
\title{Plot Posterior Predictive Checks}
\usage{
\method{plot}{bdeb_ppc}(x, n_draws = 50, ...)
}
\arguments{
\item{x}{A \code{\link[=bdeb_ppc]{bdeb_ppc()}} object.}

\item{n_draws}{Number of replicated trajectories to show. Default 50.}

\item{...}{Ignored.}
}
\value{
A ggplot2 object.
}
\description{
Plot Posterior Predictive Checks
}
\examples{
\donttest{
if (requireNamespace("cmdstanr", quietly = TRUE) &&
    nzchar(tryCatch(cmdstanr::cmdstan_path(), error = function(e) ""))) {
  data(eisenia_growth)
  dat <- bdeb_data(growth = eisenia_growth[eisenia_growth$id == 1, ])
  fit <- bdeb_fit(bdeb_model(dat, type = "individual"),
                  chains = 2, iter_warmup = 200, iter_sampling = 200,
                  refresh = 0)
  plot(bdeb_ppc(fit, type = "growth"))
}
}
}
